Adding Functionality to RW Code
I didnt realise how confusing this is to look at. I will summarize as best as possible.

Ignore the imports, I worked on this file in a separate document which I should send you to be complete, but for now, I'll commit these changes to get accustomed to this process.

 `def calc_gas_props`:

The function should calculate diffusion in a self-atmosphere and a dilute gaseous scenario such as air in nitrogen or argon in nitrogen etc.
 Given the user has the bonding energy of the gases, I added the option of using the Lennard-Jones potential to calculate Dab and the mean free path associated with that.
Added Knudsen diffusion calculation and the Knudsen direction for future use, potentially with skewed vectors or just diffusion in planar images

 `def compute_steps`:

Most of the changes here stem from changing the way I calculated kinetics. I also calculated the "skewed path length."
`rw`:

I added the if statement to allow for the skewed option because of how the code is structured.
I attempted to rewrite the code for new_vectors to change this up a bit, but the code isn't as efficient for larger numbers of walkers, say 10^8.
 T, ma, v_rel_o = kwargs['T'], kwargs['ma'], kwargs['v_rel'] is the line used to generate the distribution and maxwell.rvs is something I found on scipy's website. I got most of my info from the wiki page and the references therein.
https://en.wikipedia.org/wiki/Maxwell%E2%80%93Boltzmann_distribution

`new_vector`:
I added the option for skewing.
Using np.random uniform, I'm not too sure how this does for entropy in terms of random generators, but the distribution isn't the worst for 100k steps. Im not 100% sure about this one, so I'll take any feedback here.